fix(core): renamed a file
refs. metriport/metriport-internal#1730
